% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crimes.R
\name{get_crimes}
\alias{get_crimes}
\title{get_crimes}
\usage{
get_crimes(polygon, start_month = "latest", end_month = "latest")
}
\arguments{
\item{polygon}{an sf polygon within which to request the crimes data}

\item{start_month}{first month for which to get data for in the format "YYYY-MM". Defaults to most recent month available.}

\item{end_month}{last month for which to get data for in the format "YYYY-MM". Defaults to most recent month available.}
}
\value{
street level crimes as dataframe
}
\description{
Fetches crimes from the police.uk API within a given sf polygon.
}
\examples{
camden_poly <- localauth_ldn[localauth_ldn$lad20nm == 'Camden',]
get_crimes(camden_poly, start_month = '2020-03', end_month = '2020-07')

}
